pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "google-genai>=1.27.0",
    "ipykernel>=6.30.0",
    "langchain>=0.3.27",
    "langchain-gemini>=0.1.1",
    "langchain-google-genai>=2.1.8",
    "langchain-groq>=0.3.6",
    "langchain-ollama>=0.3.6",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
